add:settings edit

+++ src/services/apiCabins.js
@@ -18,28 +18,38 @@ export async function deleteCabin(id) {
   return data;
 }
 
-export async function createCabin(newCabin) {
+export async function createEditCabin(newCabin, id) {
+  const hasImagePath = newCabin.image?.startsWith?.(supabaseUrl);
   const imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll(
     "/",
     ""
   );
-  const imagePath = `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
-  //https://mcnwhggudkiuzfbxuaiu.supabase.co/storage/v1/object/public/cabin-images/cabin-001.jpg
-  const { data, error } = await supabase
-    .from("cabins")
-    .insert([{ ...newCabin, image: imagePath }]);
+  const imagePath = hasImagePath
+    ? newCabin.image
+    : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
+
+  //1 create/edit cabins
+  let query = supabase.from("cabins");
+
+  //create
+  if (!id) query = query.insert([{ ...newCabin, image: imagePath }]);
+  //update
+  if (id) query = query.update({ ...newCabin, image: imagePath }).eq("id", id);
+
+  const { data, error } = await query.select().single();
   if (error) {
     console.log(error);
     throw new Error("Cabin could not be deleted");
   }
   //Upload Image
+  if (hasImagePath) return data;
+
   const { error: storageError } = await supabase.storage
     .from("cabin-images")
     .upload(imageName, newCabin.image);
   //Delete the cabin if error with uploading image
   if (storageError) {
     await supabase.from("cabins").delete().eq("id", data.id);
-    console.log(storageError);
     throw new Error(
       "Cabin image could not be uploaded and he cabin was not created"
     );
